Remove protocol extension
Caused a crash when checking out with Apple Pay which is fixed with this
